fix: update paths and order after test move
