+++ src/ts/tests/noteViewPhrase.test.ts
