+++ src/ts/tests/phrase.test.ts
@@ -392,3 +392,18 @@ test('constructor fills missing trajectory/chikari grids when undefined', () => 
   expect(phrase.chikariGrid[1]).toEqual({});
   expect(phrase.chikariGrid[2]).toEqual({});
 });
+import { expect, test } from 'vitest';
+import { Phrase, initPhraseCategorization } from '../model';
+
+// Ensure missing "Bol Alap" property gets initialized to false
+
+test('constructor fills missing Bol Alap categorization', () => {
+  const custom = initPhraseCategorization();
+  // Remove the property so constructor must add it
+  delete custom.Elaboration['Bol Alap'];
+
+  const phrase = new Phrase({ categorizationGrid: [custom] });
+
+  expect(phrase.categorizationGrid[0].Elaboration['Bol Alap']).toBe(false);
+});
+
