--- src/ts/tests/trajectory-names.test.ts
@@ -1,8 +1,0 @@
-import { expect, test } from 'vitest';
-import { Trajectory } from '../model';
-
-test('Trajectory.names matches instance names', () => {
-  const staticNames = Trajectory.names();
-  const instance = new Trajectory();
-  expect(staticNames).toEqual(instance.names);
-});
